# Properly sizes the buffer when a fixed-size stream is provided to the incremental reader with default configuration.